+++ Java How to program/SolarSystemSizing/src/solarsystemsizing/TPower.java
@@ -68,7 +68,7 @@ public class TPower extends Total{
     }
     
     @Override
-    public double total() {
-        return super.getVal() * returnTEnergy();
+    public float total() {
+        return (float) (super.getVal() * returnTEnergy());
     }
 }
