+++ Java How to program/SolarSystemSizing/src/solarsystemsizing/Total.java
@@ -15,5 +15,5 @@ public abstract class Total {
         return val;
     }
     
-    public abstract double total();
+    public abstract float total();
 }
